requirements.txt
# Core ML & Data Science (Python 3.12 compatible)
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0

# Deep Learning & NLP (M1/M2 Mac compatible)
torch>=2.2.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Data & Utilities
requests>=2.31.0
python-dotenv>=1.0.0
aiofiles>=23.2.0

# SMS Gateway
twilio>=9.0.0

# Model Monitoring
evidently>=0.4.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Jupyter (optional, for analysis)
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Code Quality
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
